fix(jellyfin): use docker image to run
Figured out how to setup docker-containers which runs them as systemd
services. I had to do it because the dotnet-sdk package didn't work on
my rockpro for whatever reason because it's mislabeled as aarch64 and
the package didn't like the label. Probably could've just fixed that but
thought this might be a better solution for non-nixos stuff.